requirements.txt
wheel
Django>=5
pyTelegramBotAPI==4.26.0
python-dotenv==1.0.1
python-telegram-bot==13.15
urllib3==1.26.15
standard-imghdr==3.13.0
